refactored smoke tests in notebooks
