+++ tests/integration/test_notebooks_python.py
@@ -29,3 +29,33 @@ def test_sar_single_node_integration(notebooks, size, expected_values):
 
     for key, value in expected_values.items():
         assert results[key] == pytest.approx(value, rel=TOL)
+
+
+@pytest.mark.integration
+@pytest.mark.parametrize(
+    "size, expected_values",
+    [
+        ("1m", dict(rmse=0.89,
+                    mae=0.70,
+                    rsquared=0.36,
+                    exp_var=0.36,
+                    map=0.011,
+                    ndcg=0.10,
+                    precision=0.093,
+                    recall=0.025)),
+        # 10m works but takes too long
+    ],
+)
+def test_surprise_svd_integration(notebooks, size, expected_values):
+    notebook_path = notebooks["surprise_svd_deep_dive"]
+    pm.execute_notebook(
+        notebook_path,
+        OUTPUT_NOTEBOOK,
+        kernel_name=KERNEL_NAME,
+        parameters=dict(MOVIELENS_DATA_SIZE=size),
+    )
+    results = pm.read_notebook(OUTPUT_NOTEBOOK).dataframe.set_index("name")["value"]
+
+    for key, value in expected_values.items():
+        assert results[key] == pytest.approx(value, rel=TOL)
+
