--- tests/smoke/test_template_gpu.py
@@ -1,12 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-
-import pytest
-from reco_utils.common.gpu_utils import get_number_gpus
-
-
-@pytest.mark.smoke
-@pytest.mark.gpu
-def test_gpu_vm():
-    assert get_number_gpus() >= 1
-
